requirements.txt
PyQt6
Pillow
google-generativeai
reportlab
python-docx

